Refactor: consolidate all neural-net code under networks/
Group the model architecture code that was split between modules/ and the
vendored ocr_lib/ into a single networks/ package:

- networks/{bit_backbone,crp_models,siamese_models}.py  (moved from modules/)
- networks/ocr/                                          (moved from ocr_lib/)

modules/ now holds only the pipeline components (awl_detector, crp_classifier,
crp_locator, logo_matching); models/ remains the generated weights/data folder,
so 'models' (artifacts) and 'networks' (code) are no longer conflated.

Updated all imports plus CI (coverage include/omit, ruff exclude), SonarCloud
exclusions, and the README project structure.
