+++ .github/workflows/setup.yml
@@ -111,7 +111,7 @@ jobs:
 
       - name: Run pytest
         run: |
-          pixi run coverage run --include="modules/*.py, configs.py, phishintention.py" --omit="ocr_lib/*" -m unittest unittest_code/test_all_steps.py
+          pixi run coverage run --include="modules/*.py, networks/*.py, configs.py, phishintention.py" --omit="networks/ocr/*" -m unittest unittest_code/test_all_steps.py
           pixi run coverage xml -o coverage.xml
 
       - name: SonarQube Scan
