+++ networks/ocr/.DS_Store
