+++ networks/ocr/utils/labelmaps.py
